requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.0
bcrypt==4.1.2
pydantic==2.6.0
python-dotenv==1.0.1

